+++ sdk/keyvault/keyvault-keys/test/internal/serviceVersionParameter.spec.ts
@@ -1,19 +1,16 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
-
-import { assert } from "@azure-tools/test-utils";
-import { SinonSandbox, SinonSpy, createSandbox } from "sinon";
-import { KeyClient } from "../../src";
-import { LATEST_API_VERSION } from "../../src/keysModels";
+import { KeyClient } from "../../src/index.js";
+import { LATEST_API_VERSION } from "../../src/keysModels.js";
 import {
   HttpClient,
   PipelineRequest,
   PipelineResponse,
+  SendRequest,
   createHttpHeaders,
 } from "@azure/core-rest-pipeline";
 import { ClientSecretCredential } from "@azure/identity";
-import { versionsToTest } from "@azure-tools/test-utils";
-import { serviceVersions } from "../public/utils/common";
+import { describe, it, assert, expect, vi, beforeEach, afterEach, MockInstance } from "vitest";
 
 describe("The Keys client should set the serviceVersion", () => {
   const keyVaultUrl = `https://keyvaultname.vault.azure.net`;
@@ -33,18 +30,16 @@ describe("The Keys client should set the serviceVersion", () => {
     },
   };
 
-  let sandbox: SinonSandbox;
-  let spy: SinonSpy<[PipelineRequest], Promise<PipelineResponse>>;
+  let spy: MockInstance<SendRequest>;
   let credential: ClientSecretCredential;
   beforeEach(async () => {
-    sandbox = createSandbox();
-    spy = sandbox.spy(mockHttpClient, "sendRequest");
+    spy = vi.spyOn(mockHttpClient, "sendRequest");
 
     credential = new ClientSecretCredential("tenant", "client", "secret");
   });
 
   afterEach(() => {
-    sandbox.restore();
+    vi.restoreAllMocks();
   });
 
   it("it should default to the latest API version", async function () {
@@ -53,27 +48,24 @@ describe("The Keys client should set the serviceVersion", () => {
     });
     await client.createKey("keyName", "RSA");
 
-    const calls = spy.getCalls();
-    assert.equal(
-      calls[0].args[0].url,
-      `https://keyvaultname.vault.azure.net/keys/keyName/create?api-version=${LATEST_API_VERSION}`,
+    expect(spy).toHaveBeenCalledWith(
+      expect.objectContaining({
+        url: `https://keyvaultname.vault.azure.net/keys/keyName/create?api-version=${LATEST_API_VERSION}`,
+      }),
     );
   });
 
-  versionsToTest(serviceVersions, {}, (serviceVersion) => {
-    it("it should allow us to specify an API version from a specific set of versions", async function () {
-      const client = new KeyClient(keyVaultUrl, credential, {
-        serviceVersion: serviceVersion,
-        httpClient: mockHttpClient,
-      });
-      await client.createKey("keyName", "RSA");
-
-      const calls = spy.getCalls();
-      const lastCall = calls[calls.length - 1];
-      assert.equal(
-        lastCall.args[0].url,
-        `https://keyvaultname.vault.azure.net/keys/keyName/create?api-version=${serviceVersion}`,
-      );
+  it("it should allow us to specify an API version from a specific set of versions", async function () {
+    const client = new KeyClient(keyVaultUrl, credential, {
+      serviceVersion: "7.0",
+      httpClient: mockHttpClient,
     });
+    await client.createKey("keyName", "RSA");
+
+    expect(spy).toHaveBeenCalledWith(
+      expect.objectContaining({
+        url: `https://keyvaultname.vault.azure.net/keys/keyName/create?api-version=7.0`,
+      }),
+    );
   });
 });