+++ sdk/monitor/monitor-opentelemetry-exporter/CHANGELOG.md
@@ -7,6 +7,7 @@
 - Enforce property length limits on telemetry using truncation.
 - Updated OTel dependencies.
 - Update Statsbeat shutdown logic to include more status codes for shutdown.
+- Add non-essential statsbeat metrics.
 
 ## 1.0.0-beta.25 (2024-08-14)